move socket to its own file & fix reading
